requirements.txt
langchain
pypdf
pymupdf
pdfplumber
pytesseract
pillow
langchain_community 
faiss-cpu 
sentence-transformers
python-dotenv
openai